WINDUP-1724 Reduce list {source,target,tag} commands execution time: enhanced
