+++ bootstrap/src/main/java/org/jboss/windup/bootstrap/commands/windup/ListSourceTechnologiesCommand.java
@@ -1,16 +1,15 @@
 package org.jboss.windup.bootstrap.commands.windup;
 
-import org.jboss.windup.bootstrap.commands.AbstractListCommand;
+import org.jboss.windup.bootstrap.commands.AbstractListCommandWithoutFurnace;
 import org.jboss.windup.bootstrap.commands.Command;
 import org.jboss.windup.bootstrap.commands.CommandPhase;
 import org.jboss.windup.bootstrap.commands.CommandResult;
-import org.jboss.windup.bootstrap.commands.FurnaceDependent;
 import org.jboss.windup.exec.configuration.options.SourceOption;
 
 /**
  * @author <a href="mailto:lincolnbaxter@gmail.com">Lincoln Baxter, III</a>
  */
-public class ListSourceTechnologiesCommand extends AbstractListCommand implements Command, FurnaceDependent
+public class ListSourceTechnologiesCommand extends AbstractListCommandWithoutFurnace implements Command
 {
     @Override
     public CommandResult execute()
@@ -22,6 +21,6 @@ public class ListSourceTechnologiesCommand extends AbstractListCommand implement
     @Override
     public CommandPhase getPhase()
     {
-        return CommandPhase.PRE_EXECUTION;
+        return CommandPhase.PRE_CONFIGURATION;
     }
 }
