+++ bootstrap/src/main/java/org/jboss/windup/bootstrap/commands/windup/ListTagsCommand.java
@@ -1,16 +1,15 @@
 package org.jboss.windup.bootstrap.commands.windup;
 
-import org.jboss.windup.bootstrap.commands.AbstractListCommand;
+import org.jboss.windup.bootstrap.commands.AbstractListCommandWithoutFurnace;
 import org.jboss.windup.bootstrap.commands.Command;
 import org.jboss.windup.bootstrap.commands.CommandPhase;
 import org.jboss.windup.bootstrap.commands.CommandResult;
-import org.jboss.windup.bootstrap.commands.FurnaceDependent;
 import org.jboss.windup.exec.configuration.options.IncludeTagsOption;
 
 /**
  * @author <a href="mailto:lincolnbaxter@gmail.com">Lincoln Baxter, III</a>
  */
-public class ListTagsCommand extends AbstractListCommand implements Command, FurnaceDependent
+public class ListTagsCommand extends AbstractListCommandWithoutFurnace implements Command
 {
     @Override
     public CommandResult execute()
@@ -22,7 +21,7 @@ public class ListTagsCommand extends AbstractListCommand implements Command, Fur
     @Override
     public CommandPhase getPhase()
     {
-        return CommandPhase.PRE_EXECUTION;
+        return CommandPhase.PRE_CONFIGURATION;
     }
 
 }
